+++ src/demo/demo.h
@@ -67,15 +67,6 @@ TB_DEMO_MAIN_DECL(libm_float);
 TB_DEMO_MAIN_DECL(libm_double);
 TB_DEMO_MAIN_DECL(libm_integer);
 
-// asio
-TB_DEMO_MAIN_DECL(asio_dns);
-TB_DEMO_MAIN_DECL(asio_http);
-TB_DEMO_MAIN_DECL(asio_httpd);
-TB_DEMO_MAIN_DECL(asio_aiopc);
-TB_DEMO_MAIN_DECL(asio_aiopd);
-TB_DEMO_MAIN_DECL(asio_aicpc);
-TB_DEMO_MAIN_DECL(asio_aicpd);
-
 // math
 TB_DEMO_MAIN_DECL(math_fixed);
 TB_DEMO_MAIN_DECL(math_random);