+++ src/tbox/platform/platform.h
@@ -63,8 +63,5 @@
 #include "environment.h"
 #include "thread_pool.h"
 #include "thread_local.h"
-#ifdef TB_CONFIG_API_HAVE_DEPRECATED
-#   include "deprecated/deprecated.h"
-#endif
 
 #endif