+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Move docs directory to tbox-docs repo
 * Support tinyc compiler
+* Remove deprecated module (asio), please use coroutine module
 
 ### Bugs fixed
 
@@ -156,6 +157,7 @@
 
 * 移除docs目录，放置到独立tbox-docs仓库，减少tbox.zip包大小
 * 支持tinyc编译器
+* 移除被废弃的模块（asio模块，先用coroutine代替）
 
 ### Bugs修复